Move HTML tree-printing to its own file
Plus a few refactoring cleanups that don't change the observed behavior:

* Extract `add_children` so that `print_tree` isn't quite so huge a
  method
* Use `chars().any(...)` instead of `trim`ming and checking for an empty
  string. This is clearer and should be faster, too, since `any` can
  short-circuit but `trim` can't.
